AddScoped() -
AddScoped() - This method creates a Scoped service. A new instance of a Scoped service is created once per request within the scope. For example, in a web application it creates 1 instance per each http request but uses the same instance in the other calls within that same web request.
Issue a request to http://localhost:/home/create. 
We see the Total Employees Count is 3.
Provide a Name and click the Create button.
The Total Employees Count increases to 4.
When we click the Create button again, the Total Employees Count is still 4.